casting on size_t, to be reviewed

+++ src/ast/ast.c
@@ -6,7 +6,7 @@
 /*   By: ishenriq <ishenriq@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/19 14:31:38 by ishenriq          #+#    #+#             */
-/*   Updated: 2024/07/01 19:34:32 by paranha          ###   ########.org.br   */
+/*   Updated: 2024/07/03 12:46:48 by paranha          ###   ########.org.br   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ static int	ft_branch(t_vector *phrase, size_t pos, t_node *root)
 {
 	t_vector	*right;
 
-	if (!phrase || pos < 0 || pos > phrase->size || !root)
+	if (!phrase || (long)pos < 0 || pos > phrase->size || !root)
 		return (0);
 	right = ft_vector_slice_right(phrase, pos);
 	root->type = ft_value_int(phrase, pos, 1);
